+++ TexTransCore/LayerData/AbstractLayerData.cs
@@ -2,7 +2,7 @@ using System;
 using net.rs64.TexTransCore.BlendTexture;
 using UnityEngine;
 
-namespace net.rs64.TexTransCore.Layer
+namespace net.rs64.TexTransCore.LayerData
 {
 
     [Serializable]
