+++ TexTransCore/LayerData/AbstractLayerData.cs.meta
