+++ TexTransCore/LayerData/LayerFolderData.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using UnityEngine;
 
-namespace net.rs64.TexTransCore.Layer
+namespace net.rs64.TexTransCore.LayerData
 {
     [Serializable]
     public class LayerFolderData : AbstractLayerData
