+++ TexTransCore/LayerData/LayerFolderData.cs.meta
