+++ TexTransCore/LayerData/RasterLayerData.cs.meta
